+++ pylatex/numpy.py
@@ -1,7 +1,0 @@
-# flake8: noqa
-
-"""
-Stub numpy module for backwards compatibility.
-"""
-
-from . import Matrix, VectorName